requirements.txt
pyyaml
loguru
python-hcl2
pytest==8.0.0